Api/requirements.txt
mysql-connector-python==8.1.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
Flask-Cors==4.0.0